Make the factory return a `HubspotClient` that implements a `HubspotClientInterface`
