+++ tests/Integration/Resources/ObjectPropertyGroupsTest.php
@@ -2,7 +2,7 @@
 
 namespace SevenShores\Hubspot\Tests\Integration\Resources;
 
-use SevenShores\Hubspot\Factory;
+use SevenShores\Hubspot\HubspotClientFactory;
 use SevenShores\Hubspot\Tests\Integration\Abstraction\PropertyGroupsTestCase;
 
 /**
@@ -23,7 +23,7 @@ class ObjectPropertyGroupsTest extends PropertyGroupsTestCase
 
     public function setUp()
     {
-        $this->resource = Factory::create(getenv('HUBSPOT_TEST_API_KEY'))->objectProperties('products');
+        $this->resource = HubspotClientFactory::create(getenv('HUBSPOT_TEST_API_KEY'))->objectProperties('products');
         sleep(1);
         $this->entity = $this->createEntity();
     }
